pyproject.toml
[project]
name = "filmspot"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
	"torch",
	"torchvision",
	"transformers",
	"openai-clip",
	"openai-whisper",
	"Pillow",
	"ffmpeg-python",
	"networkx",
	"scikit-learn",
	"numpy",
	"chromadb",
	"fastapi",
	"uvicorn",
	"gradio",
	"google-generativeai",
	"pydantic",
	"rank-bm25",
	"tqdm",
]
